refine c and go internal file to _warp
